Rename algorithms group to sorting
